+++ data.lua
@@ -1,5 +1,3 @@
-require("prototypes.autoplace-controls")
-require("prototypes.resources")
-require("prototypes.recipes")
-require("prototypes.items")
-require("prototypes.technologies")
+require("prototypes.ores.autoplace-controls")
+require("prototypes.ores.quartz.resource")
+require("prototypes.ores.quartz.item")
