+++ prototypes/ores/autoplace-controls.lua
